+++ contracts/p0/BasketHandler.sol
@@ -148,7 +148,7 @@ contract BasketHandlerP0 is ComponentP0, IBasketHandler {
     // A history of baskets by basket nonce; includes current basket
     mapping(uint48 => Basket) private basketHistory;
 
-    bool public revaluable; // whether the total weights of the target basket can be changed
+    bool public reweightable; // whether the total weights of the target basket can be changed
 
     // ==== Invariants ====
     // basket is a valid Basket:
@@ -170,7 +170,7 @@ contract BasketHandlerP0 is ComponentP0, IBasketHandler {
         lastStatusTimestamp = uint48(block.timestamp);
 
         disabled = true;
-        // revaluable = false;
+        // reweightable = false;
     }
 
     /// Disable the basket in order to schedule a basket refresh
@@ -248,7 +248,7 @@ contract BasketHandlerP0 is ComponentP0, IBasketHandler {
         requireValidCollArray(erc20s);
 
         // If this isn't initial setup, require targets remain constant
-        if (!revaluable && config.erc20s.length > 0) {
+        if (!reweightable && config.erc20s.length > 0) {
             requireConstantConfigTargets(erc20s, targetAmts);
         }
 
@@ -572,9 +572,9 @@ contract BasketHandlerP0 is ComponentP0, IBasketHandler {
     }
 
     /// @custom:governance
-    function setRevaluable(bool val) public governance {
-        emit RevaluableChanged(revaluable, val);
-        revaluable = val;
+    function setReweightable(bool val) public governance {
+        emit ReweightableChanged(reweightable, val);
+        reweightable = val;
     }
 
     /* _switchBasket computes basket' from three inputs: